backend/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.2
Werkzeug==2.2.3
python-dotenv==1.0.0
pytest==7.3.1
